Prevent duplicate terms
We find boost terms for the original search query terms and for
synonyms. If the same boost terms are configured for two or more
synonyms, the terms would be added to the boost terms multiple times.
This change converts the terms to Strings and adds them to a Set to
prevent duplicates. The goal of this change is to reduce the size of the
resulting boost query.