Use bmax.inspect for the boost query
Make the boost term query a nested bmax query with locally disabled
expansion features. If bmax.inspect is true, it will be enabled for the
boost query as well and can significantly reduce the size of the parsed
boost query, because it will not include term/field combinations which
are known to never occur in the index.